requirements.txt
flowlauncher
python-dotenv
pyperclip
send2trash
